# services/events-service/python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Events API for ZeroWing Event Management System."
  version: "1.0.0"
  title: "Events API"
host: "localhost:8081"
basePath: "/v1"
schemes:
- "https"
- "http"
paths:
  /events:
    get:
      tags:
      - "Events"
      summary: "Retrieve a List of Events. Search by Event Title, Event category or\
        \ Event Description."
      operationId: "list_events"
      produces:
      - "application/json"
      parameters:
      - name: "event_title"
        in: "query"
        description: "The Event Title to search for."
        required: false
        type: "string"
      - name: "event_category"
        in: "query"
        description: "The Event Category to search for."
        required: false
        type: "string"
      - name: "event_desc"
        in: "query"
        description: "The Event Description to search for."
        required: false
        type: "string"
      - name: "host_id"
        in: "query"
        description: "The Host ID to search for."
        required: false
        type: "string"
      - name: "event_status"
        in: "query"
        description: "The Event Status to search for."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Events. Returns empty array if no Events\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/EventList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    post:
      tags:
      - "Events"
      summary: "Used to create an Event."
      operationId: "create_event"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Event object containing the Event details."
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "201":
          description: "Event created."
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
  /events/{event_id}:
    get:
      tags:
      - "Events"
      summary: "Retrieve Event details by Event ID."
      operationId: "get_event_details"
      produces:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be retrieved."
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "Event details successfully retrieved."
          schema:
            $ref: "#/definitions/Event"
        "404":
          description: "Event not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    put:
      tags:
      - "Events"
      summary: "Used to update the Event details. Replaces the Event resource."
      operationId: "update_event"
      consumes:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Event object to update. Performs a complete replace of the Event\
          \ details."
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        "200":
          description: "Event successfully updated."
          schema:
            $ref: "#/definitions/Event"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "404":
          description: "Event not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
    patch:
      tags:
      - "Events"
      summary: "Used to update the Status of a single Event e.g. Cancel an Event."
      operationId: "update_event_status"
      consumes:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "path"
        description: "ID of the Event to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "The patch operation to perform. Only Event status update is\
          \ supported."
        required: true
        schema:
          $ref: "#/definitions/EventStatusUpdate"
      responses:
        "200":
          description: "Event status successfully updated."
          schema:
            $ref: "#/definitions/Event"
        "404":
          description: "Timeslot not found."
          schema:
            $ref: "#/definitions/EventNotFoundError"
      x-swagger-router-controller: "swagger_server.controllers.events_controller"
...
definitions:
...
  EventList:
    type: "array"
    items:
      $ref: "#/definitions/Event"
...
  Event:
    type: "object"
    properties:
      event_id:
        type: "integer"
        format: "int64"
        example: 1
      host_id:
        type: "integer"
        format: "int64"
        example: 12
      account_id:
        type: "integer"
        format: "int64"
        example: 10
      venue_id:
        type: "integer"
        format: "int64"
        example: 9
      gen_seat_price:
        type: "number"
        format: "float"
        example: 100.5
      front_seat_price:
        type: "number"
        format: "float"
        example: 80.9
      mid_seat_price:
        type: "number"
        format: "float"
        example: 70.9
      back_seat_price:
        type: "number"
        format: "float"
        example: 50.0
      event_title:
        type: "string"
        example: "Sydney KPOP Party"
      event_category:
        type: "string"
        example: "Music"
      event_short_desc:
        type: "string"
        example: "Sydney KPOP Party BTS Special!"
      event_desc:
        type: "string"
        example: "STRICTLY KPOP & K-HIPHOP! KPOP ALBUM GIVEAWAYS! LIVE DJS!"
      event_start_datetime:
        type: "string"
        example: "2015-03-25T19:00:00+10:00"
      event_end_datetime:
        type: "string"
        example: "2015-03-25T23:00:00+10:00"
      event_location:
        type: "string"
        example: "Shark Hotel Sydney, NSW"
      event_img:
        type: "string"
        example: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      event_status:
        type: "string"
        enum:
        - "Upcoming"
        - "Cancelled"
        - "Completed"
      tags:
        type: "array"
        items:
          $ref: "#/definitions/Tag"
    example:
      gen_seat_price: 100.5
      front_seat_price: 80.9
      event_end_datetime: "2015-03-25T23:00:00+10:00"
      event_location: "Shark Hotel Sydney, NSW"
      event_title: "Sydney KPOP Party"
      mid_seat_price: 70.9
      event_category: "Music"
      event_start_datetime: "2015-03-25T19:00:00+10:00"
      host_id: 12
      tags:
      - name: "Jazz"
      - name: "Jazz"
      event_short_desc: "Sydney KPOP Party BTS Special!"
      event_desc: "STRICTLY KPOP & K-HIPHOP! KPOP ALBUM GIVEAWAYS! LIVE DJS!"
      event_id: 1
      account_id: 10
      back_seat_price: 50.0
      event_img: "1603dfd6-efb6-11ec-8ea0-0242ac120002.jpeg"
      venue_id: 9
      event_status: "Upcoming"
  Tag:
    type: "object"
    properties:
      name:
        type: "string"
        example: "Jazz"
    example:
      name: "Jazz"
...
  EventStatusUpdate:
    type: "object"
    required:
    - "op"
    - "path"
    - "value"
    properties:
      op:
        type: "string"
        example: "replace"
        enum:
        - "replace"
      path:
        type: "string"
        enum:
        - "/event_status"
      value:
        type: "string"
        example: "Cancelled"
        enum:
        - "Upcoming"
        - "Completed"
        - "Cancelled"
    example:
      op: "replace"
      path: "/event_status"
      value: "Cancelled"
...
  Error:
    type: "object"
    properties:
      code:
        type: "string"
      type:
        type: "string"
      message:
        type: "string"
  EventNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "EventNotFoundError"
      message: "The following Event does not exist: %event_id%"
...
  InvalidInputError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 400
      type: "InvalidInputError"
      message: "The following mandatory fields were not provided: %field_name_1%,\
        \ %field_name_2%"
  UnexpectedServiceError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 500
      type: "UnexpectedServiceError"
      message: "The following unexpected runtime error has occurred: %error%"